backend/requirements.txt
git+git://github.com/django/django@1.7c2
django-extensions==1.3.10
djangorestframework==2.3.14
gnureadline==6.3.3
ipython==2.2.0
nodeenv==0.11.0
six==1.7.3
